frontend/school_agent_project/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
redis
gunicorn
